Making a digital stethoscope that receives lung and heart signals
In this project, a digital stethoscope that detects and processes heart and lung sounds is being developed. The design includes high and low pass filters, amplification, offset addition, analog to digital converter. The processed data is sent to MATLAB using Arduino and processed there. Finally, heart and lung frequencies can be filtered in MATLAB and the disease can be diagnosed. The project consists of two parts: software and hardware. This was done by Mustafa ÇİVDEM, Zeynep DOLGUN and Rabia ERDEM as a sophomore term project.